Clean-up_first - first attempt

+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,6 +1,5 @@
 package com.kodilla.ecommercee.controller;
 
-import com.kodilla.ecommercee.domain.Group;
 import com.kodilla.ecommercee.domain.GroupDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,8 +10,9 @@ import java.util.List;
 @RestController
 @RequestMapping("/groups")
 public class GroupController {
+
     private final List<GroupDto> productGroups = new ArrayList<>();
-    private int nextGroupId =1;
+    private Long nextGroupId = 1L;
 
     @GetMapping
     public List<GroupDto> getAllGroups() {
@@ -36,8 +36,9 @@ public class GroupController {
         }
         return ResponseEntity.notFound().build();
     }
+
     @PutMapping("/{id}")
-    public ResponseEntity<GroupDto> updateGroup(@PathVariable("id") int id, @RequestBody GroupDto updatedGroup) {
+    public ResponseEntity<GroupDto> updateGroup(@PathVariable("id") Long id, @RequestBody GroupDto updatedGroup) {
         for (GroupDto group : productGroups) {
             if (group.getId() == id) {
                 updatedGroup.setId(id);
@@ -49,4 +50,3 @@ public class GroupController {
         return ResponseEntity.notFound().build();
     }
 }
-
